Google Advanced Data Analytics Capstone Project
This project was a capstone for my course, and here I have wrangled the data and moved things around, curiously created visuals about the data. At the end,  I trained a RandomForestClassifier model , fed the data, feature engineered few variables, and checked the prediction results to evaluate the model's performance. The model performed very well, and the result was satisfying.